- Blokiranje, aktiviranje i brisanje korisnika - Provera pristupa

+++ administrator/src/main/java/com/xml/administrator/controllers/CodebookController.java
@@ -1,5 +1,9 @@
 package com.xml.administrator.controllers;
 
+import java.util.ArrayList;
+
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -10,8 +14,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.xml.administrator.dto.AgentDTO;
 import com.xml.administrator.dto.ContentDTO;
+import com.xml.administrator.model.Admin;
 import com.xml.administrator.model.ObjectCategory;
 import com.xml.administrator.model.ObjectType;
 import com.xml.administrator.model.Services;
@@ -25,14 +29,27 @@ public class CodebookController {
 	@Autowired
 	private CodebookService codeSer;
 	
+	@Autowired
+	private HttpServletRequest request;
+	
 	// ----- Object Category
 	@RequestMapping(value ="/objCategory/all", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity getObjCategories() {
+	public ResponseEntity<ArrayList<ObjectCategory>> getObjCategories() {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		return ResponseEntity.status(200).body(codeSer.getAllObjCategories());
 	}
 	
 	@RequestMapping(value = "/objCategory/add", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity addObjCategory(@RequestBody ContentDTO content) throws Exception {
+	public ResponseEntity<ObjectCategory> addObjCategory(@RequestBody ContentDTO content) throws Exception {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		ObjectCategory temp = codeSer.addObjCate(content);
 		if (temp != null) {
 			return new ResponseEntity<ObjectCategory>(temp, HttpStatus.CREATED);
@@ -42,6 +59,11 @@ public class CodebookController {
 	
 	@RequestMapping(value = "/objCategory/{id}", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<ObjectCategory> editObjCategory(@PathVariable  Long id, @RequestBody ContentDTO content) throws Exception {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		ObjectCategory temp = codeSer.editObjCate(id, content.getContent());
 		if (temp == null) {
 			return new ResponseEntity<ObjectCategory>(temp, HttpStatus.BAD_REQUEST);
@@ -49,8 +71,13 @@ public class CodebookController {
 		return new ResponseEntity<ObjectCategory>(temp, HttpStatus.OK);
 	}
 	
-	@RequestMapping(value = "/objCategory/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
+	@RequestMapping(value = "/objCategory/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<String> deleteObjCategory(@PathVariable  Long id) throws Exception {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity<String>("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		ObjectCategory temp = codeSer.deleteObjCate(id);
 		if (temp == null) {
 			return new ResponseEntity<String>("Nije izbrisano.", HttpStatus.BAD_REQUEST);
@@ -61,12 +88,22 @@ public class CodebookController {
 	// ---- Object type
 	
 	@RequestMapping(value ="/objType/all", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity getObjTypes() {
+	public ResponseEntity<ArrayList<ObjectType>> getObjTypes() {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		return ResponseEntity.status(200).body(codeSer.getAllObjTypes());
 	}
 	
 	@RequestMapping(value = "/objType/add", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<ObjectType> addObjType(@RequestBody ContentDTO content) throws Exception {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		ObjectType temp = codeSer.addObjType(content);
 		if (temp != null) {
 			return new ResponseEntity<ObjectType>(temp, HttpStatus.CREATED);
@@ -76,6 +113,11 @@ public class CodebookController {
 	
 	@RequestMapping(value = "/objType/{id}", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<ObjectType> editObjType(@PathVariable  Long id, @RequestBody ContentDTO content) throws Exception {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		ObjectType temp = codeSer.editObjType(id, content.getContent());
 		if (temp == null) {
 			return new ResponseEntity<ObjectType>(temp, HttpStatus.BAD_REQUEST);
@@ -83,8 +125,13 @@ public class CodebookController {
 		return new ResponseEntity<ObjectType>(temp, HttpStatus.OK);
 	}
 	
-	@RequestMapping(value = "/objType/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
+	@RequestMapping(value = "/objType/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<String> deleteObjType(@PathVariable  Long id) throws Exception {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity<String>("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		ObjectType temp = codeSer.deleteObjType(id);
 		if (temp == null) {
 			return new ResponseEntity<String>("Nije izbrisano.", HttpStatus.BAD_REQUEST);
@@ -95,12 +142,21 @@ public class CodebookController {
 	// ---- Services
 	
 	@RequestMapping(value ="/services/all", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity getServices() {
+	public ResponseEntity<ArrayList<Services>> getServices() {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		return ResponseEntity.status(200).body(codeSer.getAllServices());
 	}
 	
 	@RequestMapping(value = "/services/add", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity addServices(@RequestBody ContentDTO content) throws Exception {
+	public ResponseEntity<Services> addServices(@RequestBody ContentDTO content) throws Exception {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
 		Services temp = codeSer.addServ(content);
 		if (temp != null) {
 			return new ResponseEntity<Services>(temp, HttpStatus.CREATED);
@@ -110,6 +166,11 @@ public class CodebookController {
 	
 	@RequestMapping(value = "/services/{id}", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Services> editServices(@PathVariable  Long id, @RequestBody ContentDTO content) throws Exception {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		Services temp = codeSer.editServices(id, content.getContent());
 		if (temp == null) {
 			return new ResponseEntity<Services>(temp, HttpStatus.BAD_REQUEST);
@@ -117,8 +178,13 @@ public class CodebookController {
 		return new ResponseEntity<Services>(temp, HttpStatus.OK);
 	}
 	
-	@RequestMapping(value = "/services/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
+	@RequestMapping(value = "/services/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<String> deleteServices(@PathVariable  Long id) throws Exception {
+		TUser tempAdmin = (Admin)request.getSession().getAttribute("admin");
+		if(tempAdmin == null) {
+			return new ResponseEntity<String>("Niste ulogovani.", HttpStatus.UNAUTHORIZED);
+		}
+		
 		Services temp = codeSer.deleteServices(id);
 		if (temp == null) {
 			return new ResponseEntity<String>("Nije izbrisano.", HttpStatus.BAD_REQUEST);
